Fix checkstyle violations in core
Fixes all error violations in the Core project and also refactors some
inner classes inside the debug.BenchmarkScreen to standalone classes.
