+++ modules/Core/src/main/java/org/terasology/rendering/nui/layers/hud/DropItemRegion.java
